[rework] dynamic commands import
